+++ nemo/collections/tts/helpers/helpers.py
@@ -16,6 +16,9 @@ import librosa
 import matplotlib.pylab as plt
 import numpy as np
 import torch
+import time
+
+from pytorch_lightning.callbacks.base import Callback
 
 from nemo.utils import logging
 
@@ -26,6 +29,7 @@ def get_mask_from_lengths(lengths, max_len=None):
     ids = torch.arange(0, max_len, out=torch.cuda.LongTensor(max_len))
     mask = (ids < lengths.unsqueeze(1)).bool()
     return mask
+
 
 def griffin_lim(magnitudes, n_iters=50, n_fft=1024):
     """
@@ -43,6 +47,7 @@ def griffin_lim(magnitudes, n_iters=50, n_fft=1024):
         complex_spec = magnitudes * phase
         signal = librosa.istft(complex_spec)
     return signal
+
 
 def tacotron2_log_to_tb_func(
     swriter,
@@ -95,6 +100,7 @@ def tacotron2_log_to_tb_func(
             audio = griffin_lim(magnitude.T ** griffin_lim_power)
             swriter.add_audio(f"audio/{tag}_target", audio / max(np.abs(audio)), step, sample_rate=sr)
 
+
 def plot_alignment_to_numpy(alignment, info=None):
     fig, ax = plt.subplots(figsize=(6, 4))
     im = ax.imshow(alignment, aspect='auto', origin='lower', interpolation='none')
@@ -144,21 +150,16 @@ def plot_gate_outputs_to_numpy(gate_targets, gate_outputs):
     plt.close()
     return data
 
+
 def save_figure_to_numpy(fig):
     # save it to a numpy array.
     data = np.fromstring(fig.canvas.tostring_rgb(), dtype=np.uint8, sep='')
     data = data.reshape(fig.canvas.get_width_height()[::-1] + (3,))
     return data
 
+
 def waveglow_log_to_tb_func(
-    swriter,
-    tensors,
-    step,
-    tag="train",
-    n_fft=1024,
-    hop_length=256,
-    window="hann",
-    mel_fb=None,
+    swriter, tensors, step, tag="train", n_fft=1024, hop_length=256, window="hann", mel_fb=None,
 ):
     audio_pred, spec_target, mel_length = tensors
     mel_length = mel_length[0]
@@ -169,17 +170,21 @@ def waveglow_log_to_tb_func(
     if mel_fb is not None:
         mag, _ = librosa.core.magphase(
             librosa.core.stft(
-                np.nan_to_num(audio_pred[0].cpu().detach().numpy()),
-                n_fft=n_fft,
-                hop_length=hop_length,
-                window=window,
+                np.nan_to_num(audio_pred[0].cpu().detach().numpy()), n_fft=n_fft, hop_length=hop_length, window=window,
             )
         )
         mel_pred = np.matmul(mel_fb.cpu().numpy(), mag).squeeze()
         log_mel_pred = np.log(np.clip(mel_pred, a_min=1e-5, a_max=None))
         swriter.add_image(
-            f"{tag}_mel_predicted",
-            plot_spectrogram_to_numpy(log_mel_pred[:, :mel_length]),
-            step,
-            dataformats="HWC",
+            f"{tag}_mel_predicted", plot_spectrogram_to_numpy(log_mel_pred[:, :mel_length]), step, dataformats="HWC",
         )
+
+
+class LogEpochTimeCallback(Callback):
+    def on_epoch_start(self, trainer, pl_module):
+        self.epoch_start = time.time()
+
+    def on_epoch_end(self, trainer, pl_module):
+        curr_time = time.time()
+        duration = curr_time - self.epoch_start
+        trainer.logger.log_metrics({"epoch_time": duration}, step=trainer.global_step)
