+++ nemo/collections/tts/models/__init__.py
@@ -11,3 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from nemo.collections.tts.models.waveglow import WaveglowPTL
+from nemo.collections.tts.models.tacotron2 import Tacotron2PTL
